requirements.txt

asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4

certifi==2022.9.24
chardet==3.0.4
Click==7.0
colorama==0.4.3

cryptography==2.8

dbus-python==1.2.16

distro==1.4.0
distro-info===1.0
dj-database-url==1.0.0
Django==4.1.2
django-countries==7.4.1
django-heroku==0.3.1
Django-Verify-Email==2.0.3

entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
fasteners==0.14.1
future==0.18.2
h11==0.14.0
httplib2==0.14.0
idna==2.8
keyring==18.0.1

launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==1.3
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
numpy==1.23.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
paramiko==2.6.0
pbr==5.11.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
PySocks==1.7.1
python-apt==0.0
python-dateutil==2.7.3
python-debian==0.1.49
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.5.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.2.4

testresources==2.0.1
torch==1.12.1
torchaudio==0.12.1+cpu
torchvision==0.13.1+cpu
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.4.0

ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
wadllib==1.3.3
whitenoise==6.2.0
wsproto==1.2.0
xkit==0.0.0
